# [Codegen][GPU] Enable TileAndFuse for matmul by default