--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/test/ROCDL/config_igemm_tile_and_fuse.mlir
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/test/ROCDL/config_igemm_tile_and_fuse.mlir
@@ -31,11 +31,11 @@ func.func @nhwc_conv_mfma() {
 //  CHECK-SAME:     mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>
 //  CHECK-SAME:     promote_operands = [0, 1]
 
-//  GFX942-SAME:    reduction = [0, 0, 0, 0, 8]
+//  GFX942-SAME:    reduction = [0, 0, 0, 0, 16]
 //  GFX942-SAME:    subgroup = [1, 2, 1, 2, 0]
 //  GFX942-SAME:    workgroup = [1, 4, 32, 64, 0]
 
-//  MI300X-SAME:    reduction = [0, 0, 0, 0, 8]
+//  MI300X-SAME:    reduction = [0, 0, 0, 0, 16]
 //  MI300X-SAME:    subgroup = [1, 1, 1, 1, 0]
 //  MI300X-SAME:    workgroup = [1, 2, 32, 32, 0]}>
 
@@ -65,11 +65,11 @@ func.func @nchw_conv_mfma() {
 //  CHECK-SAME:     mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>
 //  CHECK-SAME:     promote_operands = [0, 1]
 
-// GFX942-SAME:     reduction = [0, 0, 0, 0, 8]
+// GFX942-SAME:     reduction = [0, 0, 0, 0, 16]
 // GFX942-SAME:     subgroup = [1, 2, 2, 1, 0]
 // GFX942-SAME:     workgroup = [1, 64, 4, 32, 0]
 
-// MI300X-SAME:     reduction = [0, 0, 0, 0, 8]
+// MI300X-SAME:     reduction = [0, 0, 0, 0, 16]
 // MI300X-SAME:     subgroup = [1, 1, 1, 1, 0]
 // MI300X-SAME:     workgroup = [1, 32, 2, 32, 0]
 
@@ -98,15 +98,15 @@ func.func @nhwc_conv_unaligned_mfma() {
 //       CHECK:   linalg.conv_2d_nhwc_hwcf {{.*}}lowering_config = #iree_gpu.lowering_config
 //  CHECK-SAME:     mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>
 
-// GFX942-SAME:     padding = [2, 1, 32, 64, 32]
+// GFX942-SAME:     padding = [2, 1, 32, 64, 64]
 // GFX942-SAME:     promote_operands = [0, 1]
-// GFX942-SAME:     reduction = [0, 0, 0, 0, 8]
+// GFX942-SAME:     reduction = [0, 0, 0, 0, 16]
 // GFX942-SAME:     subgroup = [2, 1, 1, 1, 0]
 // GFX942-SAME:     workgroup = [2, 1, 32, 64, 0]
 
-// MI300X-SAME:     padding = [2, 1, 32, 32, 32]
+// MI300X-SAME:     padding = [2, 1, 32, 32, 64]
 // MI300X-SAME:     promote_operands = [0, 1]
-// MI300X-SAME:     reduction = [0, 0, 0, 0, 8]
+// MI300X-SAME:     reduction = [0, 0, 0, 0, 16]
 // MI300X-SAME:     subgroup = [1, 1, 1, 1, 0]
 // MI300X-SAME:     workgroup = [2, 1, 32, 32, 0]
 
@@ -137,15 +137,15 @@ func.func @nchw_conv_unaligned_mfma() {
 //       CHECK:   linalg.conv_2d_nchw_fchw {{.*}}lowering_config = #iree_gpu.lowering_config
 //  CHECK-SAME:     mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>
 
-// GFX942-SAME:     padding = [1, 64, 4, 32, 32]
+// GFX942-SAME:     padding = [1, 64, 4, 32, 64]
 // GFX942-SAME:     promote_operands = [0, 1]
-// GFX942-SAME:     reduction = [0, 0, 0, 0, 8]
+// GFX942-SAME:     reduction = [0, 0, 0, 0, 16]
 // GFX942-SAME:     subgroup = [1, 2, 2, 1, 0]
 // GFX942-SAME:     workgroup = [1, 64, 4, 32, 0]
 
-// MI300X-SAME:     padding = [1, 32, 2, 32, 32]
+// MI300X-SAME:     padding = [1, 32, 2, 32, 64]
 // MI300X-SAME:     promote_operands = [0, 1]
-// MI300X-SAME:     reduction = [0, 0, 0, 0, 8]
+// MI300X-SAME:     reduction = [0, 0, 0, 0, 16]
 // MI300X-SAME:     subgroup = [1, 1, 1, 1, 0]
 // MI300X-SAME:     workgroup = [1, 32, 2, 32, 0]
 
@@ -214,9 +214,9 @@ func.func @conv_chwn_chwf_unaligned_batch(%arg0: tensor<16x193x129x40xbf16>, %ar
 
 //       CHECK:   linalg.generic {{.*}}lowering_config = #iree_gpu.lowering_config
 //  CHECK-SAME:     mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x16_BF16>
-//  CHECK-SAME:     padding = [16, 1, 1, 16, 64]
+//  CHECK-SAME:     padding = [16, 1, 1, 16, 128]
 //  CHECK-SAME:     promote_operands = [0, 1]
-//  CHECK-SAME:     reduction = [0, 0, 0, 0, 4]
+//  CHECK-SAME:     reduction = [0, 0, 0, 0, 8]
 //  CHECK-SAME:     subgroup = [1, 1, 1, 1, 0]
 //  CHECK-SAME:     workgroup = [16, 1, 1, 16, 0]
 
@@ -246,19 +246,19 @@ func.func @group_conv_hwgc_gfhwc_unaligned(%arg0: tensor<61x93x16x56xbf16>, %arg
 
 //       CHECK:   linalg.generic {{.*}}lowering_config = #iree_gpu.lowering_config
 // GFX942-SAME:     mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x16_BF16>
-// GFX942-SAME:     padding = [1, 32, 1, 64, 64]
+// GFX942-SAME:     padding = [1, 32, 1, 64, 128]
 // GFX942-SAME:     promote_operands = [0, 1]
-// GFX942-SAME:     reduction = [0, 0, 0, 0, 4]
+// GFX942-SAME:     reduction = [0, 0, 0, 0, 8]
 // GFX942-SAME:     subgroup = [1, 1, 0, 1, 0]
 // GFX942-SAME:     workgroup = [1, 32, 1, 64, 0]
 
-// MI300X-SAME:     padding = [1, 32, 1, 64, 64]
+// MI300X-SAME:     padding = [1, 32, 1, 64, 128]
 // MI300X-SAME:     promote_operands = [0, 1]
-// MI300X-SAME:     reduction = [0, 0, 0, 0, 4]
+// MI300X-SAME:     reduction = [0, 0, 0, 0, 8]
 // MI300X-SAME:     subgroup = [1, 1, 0, 1, 0]
 // MI300X-SAME:     workgroup = [1, 32, 1, 64, 0]
 
-// PAD-CONV-GFX942:     padding_conv = [1, 32, 1, 64, 0, 0, 64]
+// PAD-CONV-GFX942:     padding_conv = [1, 32, 1, 64, 0, 0, 128]
 
 // -----