--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/test/config_matmul.mlir
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/test/config_matmul.mlir
@@ -1,6 +1,5 @@
 // RUN: iree-opt --split-input-file --iree-gpu-test-target=gfx942 --pass-pipeline='builtin.module(linalg-specialize-generic-ops,iree-llvmgpu-select-lowering-strategy)' %s \
 // RUN:   | FileCheck %s --check-prefixes=CHECK,SPECIALIZED
-
 // RUN: iree-opt --split-input-file --iree-gpu-test-target=gfx942 --pass-pipeline='builtin.module(linalg-generalize-named-ops,iree-llvmgpu-select-lowering-strategy)' %s \
 // RUN:   | FileCheck %s --check-prefixes=CHECK,GENERALIZED
 
@@ -14,16 +13,15 @@
 !TC = tensor<32x32xf32>
 !DTC = !iree_tensor_ext.dispatch.tensor<readwrite:tensor<32x32xf32>>
 
-//      CHECK:    #iree_codegen.translation_info<pipeline = LLVMGPUVectorDistribute
-// CHECK-SAME:    workgroup_size = [256, 1, 1] subgroup_size = 64, {}>
+//      CHECK:    #iree_codegen.translation_info<pipeline = LLVMGPUTileAndFuse
+// CHECK-SAME:    workgroup_size = [256, 1, 1] subgroup_size = 64, {
 func.func @matmul_32_32_32(%arg0: !TA, %arg1: !TB, %arg2: !TC, %arg3: !DTC) {
   // Sanity check that generalize/specialize works.
   // GENERALIZED:   linalg.generic
   // SPECIALIZED:   linalg.matmul
-  //      CHECK:  #iree_gpu.lowering_config<{mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>,
-  // CHECK-SAME:  promote_operands = [0, 1], reduction = [0, 0, 32]
-  // CHECK-SAME{LITERAL}: subgroup_basis = [[2, 2, 1], [0, 1, 2]]
-  // CHECK-SAME:  workgroup = [32, 32, 0]}>
+  //      CHECK:  {lowering_config = #iree_gpu.lowering_config<{mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>,
+  // CHECK-SAME:  promote_operands = [0, 1], reduction = [0, 0, 8], subgroup = [1, 1, 0],
+  // CHECK-SAME:  workgroup = [32, 32, 0]}>}
   %0 = linalg.matmul ins(%arg0, %arg1 : !TA, !TB) outs(%arg2 : !TC) -> !TC
   iree_tensor_ext.dispatch.tensor.store %0, %arg3, offsets = [0, 0], sizes = [32, 32], strides = [1, 1] : !TC -> !DTC
   return
@@ -35,13 +33,11 @@ func.func @matmul_32_32_32(%arg0: !TA, %arg1: !TB, %arg2: !TC, %arg3: !DTC) {
 !TB = tensor<4096x4096xf32>
 !TC = tensor<4096x4096xf32>
 !DTC = !iree_tensor_ext.dispatch.tensor<readwrite:tensor<4096x4096xf32>>
-//      CHECK:    #iree_codegen.translation_info<pipeline = LLVMGPUVectorDistribute
-// CHECK-SAME:    workgroup_size = [256, 1, 1] subgroup_size = 64, {}>
+//      CHECK:    #iree_codegen.translation_info<pipeline = LLVMGPUTileAndFuse
+// CHECK-SAME:    workgroup_size = [256, 1, 1] subgroup_size = 64, {
 func.func @matmul_4096_4096_4096(%arg0: !TA, %arg1: !TB, %arg2: !TC, %arg3: !DTC) {
-  //      CHECK:  #iree_gpu.lowering_config<{mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>,
-  // CHECK-SAME:  promote_operands = [0, 1], reduction = [0, 0, 16]
-  // CHECK-SAME{LITERAL}: subgroup_basis = [[2, 2, 1], [0, 1, 2]]
-  // CHECK-SAME:  workgroup = [64, 128, 0]}>}
+  //      CHECK: {lowering_config = #iree_gpu.lowering_config<{mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>,
+  // CHECK-SAME: promote_operands = [0, 1], reduction = [0, 0, 4], subgroup = [4, 4, 0], workgroup = [128, 128, 0]}>
   %0 = linalg.matmul ins(%arg0, %arg1 : !TA, !TB) outs(%arg2 : !TC) -> !TC
   iree_tensor_ext.dispatch.tensor.store %0, %arg3, offsets = [0, 0], sizes = [4096, 4096], strides = [1, 1] : !TC -> !DTC
   return
@@ -53,12 +49,11 @@ func.func @matmul_4096_4096_4096(%arg0: !TA, %arg1: !TB, %arg2: !TC, %arg3: !DTC
 !TB = tensor<32x4096xf32>
 !TC = tensor<4096x4096xf32>
 !DTC = !iree_tensor_ext.dispatch.tensor<readwrite:tensor<4096x4096xf32>>
-//      CHECK:    #iree_codegen.translation_info<pipeline = LLVMGPUVectorDistribute
-// CHECK-SAME:    workgroup_size = [256, 1, 1] subgroup_size = 64, {}>
+//      CHECK:    #iree_codegen.translation_info<pipeline = LLVMGPUTileAndFuse
+// CHECK-SAME:    workgroup_size = [256, 1, 1] subgroup_size = 64, {
 func.func @matmul_4096_32_4096(%arg0: !TA, %arg1: !TB, %arg2: !TC, %arg3: !DTC) {
   //      CHECK:  #iree_gpu.lowering_config<{mma_kind = #iree_gpu.mma_layout<MFMA_F32_16x16x4_F32>,
-  // CHECK-SAME:  promote_operands = [0, 1], reduction = [0, 0, 16]
-  // CHECK-SAME{LITERAL}: subgroup_basis = [[2, 2, 1], [0, 1, 2]]
+  // CHECK-SAME:  promote_operands = [0, 1], reduction = [0, 0, 8], subgroup = [2, 4, 0],
   // CHECK-SAME:  workgroup = [64, 128, 0]}>}
   %0 = linalg.matmul ins(%arg0, %arg1 : !TA, !TB) outs(%arg2 : !TC) -> !TC
   iree_tensor_ext.dispatch.tensor.store %0, %arg3, offsets = [0, 0], sizes = [4096, 4096], strides = [1, 1] : !TC -> !DTC