… compiler/src/iree/compiler/Codegen/LLVMGPU/test/rocdl_pipeline_test.mlir
… tests/external/iree-test-suites/sharktank_models/benchmarks/llama/8b_f16_prefill_rocm.json
… tests/external/iree-test-suites/sharktank_models/benchmarks/sdxl/vae_rocm.json