This is the second attempt after https://github.com/iree-org/iree/pull/19854.

I'm starting this as a draft to run through CI regressions first.