@jerryyin 2025-09-03
Sharktank summary:
 - Prefill improved quite a bit from 37.8424 to 27.9952 ms
 - Clip degraded quite a bit from 7.5783 to 12.3838 ms
 - All other benchmark stay flat before and after this PR

I'll be looking at perf regression at clip.